[Port] Update to special/reindev-2.9 (v2.6.1)

+++ common/src/main/java/com/gitlab/cdagaming/craftpresence/utils/gui/controls/PresenceVisualizer.java
@@ -412,8 +412,8 @@ public class PresenceVisualizer {
         }
 
         if (lastCompiledPresence != null && childFrame.isOverScreen()) {
-            final int mouseX = screen.getMouseX();
-            final int mouseY = screen.getMouseY();
+            final float mouseX = screen.getMouseX();
+            final float mouseY = screen.getMouseY();
 
             if (smallImageData != null && RenderUtils.isMouseOver(
                     mouseX, mouseY,
